+++ CHANGELOG.md
@@ -17,6 +17,10 @@ View all releases at: <https://github.com/trimble-oss/modus-web-components/relea
 
 - Events to know when the `modus-navbar` is opening links and menus
 
+### Fixed
+
+- Fixed the thick chevron icons used in Modus Accordion and Modus Content Tree components
+
 ## 0.1.31 - 2023-01-19
 
 ### Fixed